ref. for ScatBase
Ref. for ScatBase paper, which implements an accurate lookup table method for
calculating the scattering angle for a given ion energy, screening potential,
and impact parameter.

The advantage of this approach is its speed; one avoid both the need for root
finding (to get the distance of closest approach) and numerically evaluating
the scattering integral!